pyproject.toml
[tool.poetry]
name = "ai-agent-hackason-app"
version = "0.1.0"
description = ""
authors = ["vel-3633f <vel3633fu@gmail.com>"] 
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
streamlit = ">=1.44.1,<2.0.0" 
langchain = "^0.3.24"
langchain-openai = "^0.3.14"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core>=1.0.0"] 
build-backend = "poetry.core.masonry.api"